% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_flight_plan.R
\name{assign_flight_plan}
\alias{assign_flight_plan}
\title{Assign a flight plan to a survey design}
\usage{
assign_flight_plan(
  sf_segments,
  flight_id,
  col_trans_id,
  flight_day,
  survey_start_hour,
  flight_speed,
  intertransect_gap_duration
)
}
\arguments{
\item{sf_segments}{sf object, segments to assign flight plan to}

\item{flight_id}{numeric vector, giving the order by which the segments must be surveyed}

\item{col_trans_id}{character, the column in sf_segments storing transect names}

\item{flight_day}{character, the day the survey flight is carried out (ymd format: "2022-02-08")}

\item{survey_start_hour}{character, the hour the survey is started (hms format: "06:00:00")}

\item{flight_speed}{numeric, the speed at which a single segment is surveyed (in km/h)}

\item{intertransect_gap_duration}{numeric, the duration of the transit from a transect to the other (in seconds)}
}
\value{
a sf table with four additional columns (transect_id, flight_id, start_time and end_time) informing the period over which each segment is surveyed
}
\description{
Assign a flight plan to a survey design
}
\examples{
surv <- generate_survey_plan(bbx_xmin = 5, bbx_xmax = 85, bbx_ymin = 5, bbx_ymax = 85,
                             start_x = 5, end_x = 85, start_y = 5, end_y = 85,
                             space_out_factor = 4, segmentize = TRUE, seg_length = 1, 
                             buffer = TRUE, buffer_width = 0.2
                             )

# Here, set up a flight plan where the first transect is flown in a given 
# direction, the second in the other direction etc, so that the aircraft 
# does not need to go all the way to the other side of the survey area 
# to start a new transect. The intransect.gap.duration is thus set to 1 minute, 
# a the flight speed to 160 km/h (the classical speed for aerial surveys). 
# in that case, the survey duration for a single segment is 22.5 seconds. 
flight_id <- c(1:80, 160:81, 161:240, 320:241, 321:400, 480:401, 481:560, 640:561, 641:720, 800:721, 801:880, 960:881, 961:1040,
                   1120:1041, 1121:1200, 1280:1201, 1281:1360, 1440:1361, 1441:1520, 1600:1521, 1601:1680)

# assign the time periods to each segment
flight_plan <- assign_flight_plan(
    sf_segments = surv$buffered_segments,
    flight_id = flight_id,
    col_trans_id = "transect",
    flight_day = "2022-02-08", 
    survey_start_hour = "06:00:00",
    flight_speed = 160,
    intertransect_gap_duration = 60
)

# plot to check everything is ok
library(ggplot2)
ggplot(flight_plan) +
    geom_sf(aes(fill = start_time),
            color = "black", size = 0.2) +
    scale_fill_datetime(low = "#8073ac", high = "#fdb863") +
    theme_bw() + theme(panel.grid = element_blank()) +
    labs(fill = "Hour", limits = c("06:00", "16:16"), x = "x", y = "y")+
    coord_sf(xlim = c(0,90), ylim = c(0,90))

ggplot(flight_plan) +
    geom_sf(aes(fill = flight_id),
            color = "black", size = 0.2)
}
