% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create the grid of a virtual environment}
\usage{
create_grid(
  ymin = 0,
  ymax = 90,
  xmin = 0,
  xmax = 90,
  step_x = 0.5,
  step_y = 0.5
)
}
\arguments{
\item{ymin}{Minimum latitude (default to 0)}

\item{ymax}{Maximum latitude (default to 90)}

\item{xmin}{Minimum longitude (default to 0)}

\item{xmax}{Maximum latitude (default to 90)}

\item{step_x}{Longitude steps to be used to generate points (default to 0.5)}

\item{step_y}{Latitude steps to be used to generate points (default to 0.5)}
}
\value{
A grid in dataframe format.
}
\description{
Create the grid of a virtual environment
}
\examples{
grid <- create_grid()
}
