% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_proj.R
\name{custom_proj}
\alias{custom_proj}
\title{Set up a custom projection for virtual environment}
\usage{
custom_proj(lon_0 = 0, lat_0 = 0, x_0 = 0, y_0 = 0)
}
\arguments{
\item{lon_0, lat_0}{Longitude and Latitude of the center, defaults to 0}

\item{x_0, y_0}{False easting and northing, defaults to 0}
}
\value{
A object of class crs
}
\description{
Set up a custom projection for virtual environment
}
\examples{
custom_proj()
}
