% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_functions.R
\name{detfunc_hn}
\alias{detfunc_hn}
\alias{detfunc_hr}
\title{Half-normal detection function}
\usage{
detfunc_hn(x, sigma)

detfunc_hr(x, sigma, b)
}
\arguments{
\item{x}{numeric}

\item{sigma}{numeric}

\item{b}{numeric}
}
\description{
Half-normal detection function

Hazard rate detection function
}
