% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_env_layer.R
\name{generate_env_layer}
\alias{generate_env_layer}
\title{Generating random environmental layers}
\usage{
generate_env_layer(
  seed = 2,
  grid,
  n = 2,
  beta = 1,
  norm = TRUE,
  return_rasters = TRUE
)
}
\arguments{
\item{seed}{Value to feed set.seed (default 2)}

\item{grid}{Grid of the virtual space, as generated by \code{\link{create_grid}}}

\item{n}{Number of layers to be generated (default 2)}

\item{beta}{The beta value to be forwarded to gstat (default to 1)}

\item{norm}{Should the resulting layers be normalized?}

\item{return_rasters}{Should the function returns SpatRast in addition to data frame?}
}
\value{
Returns a data frame with the simulated environmental values, as well as SpatRast is return_rasters = TRUE
}
\description{
Generating random environmental layers
}
\examples{
grid <- create_grid()

str(generate_env_layer(norm = FALSE, return_rasters = FALSE, grid = grid))

library(terra)
plot(generate_env_layer(norm = TRUE, return_rasters = TRUE, grid = grid)$rasters)
}
\seealso{
Other environment simulation functions: 
\code{\link{generate_resource_layer}()}
}
\concept{environment simulation functions}
