% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_resource_layer.R
\name{generate_resource_layer}
\alias{generate_resource_layer}
\title{Resource selection function}
\usage{
generate_resource_layer(
  beta,
  env_layers,
  coordinate_fields = c("x", "y"),
  norm = TRUE,
  return.rasters = TRUE
)
}
\arguments{
\item{beta}{The beta values to multiply the environmental layers with (must be as much beta as layers)}

\item{env_layers}{Environmental layers to build the resource from. Can either be a data frame (including map coordinates and environmental layer values) or a SpatRast.}

\item{coordinate_fields}{If env_layers is a data frame, the fields corresponding to the map coordinates (e.g. c("x", "y"), the default)}

\item{norm}{Should the resource layer be normalized?}

\item{return.rasters}{Should the function returns the resource as SpatRast in addition to the data frame?}
}
\value{
Returns a data frame with the suitability layer, as well as SpatRast is return.rasters = TRUE
}
\description{
Resource selection function
}
\examples{
library(terra)
# simple example
grid <- create_grid()
cdt <- generate_env_layer(grid = grid)
rsce <- generate_resource_layer(env_layers = cdt$rasters,
                            beta = c(2, -1.5))
str(rsce)
plot(rsce$rasters)

# also works when coordinates are not names x,y
cdt2 <- generate_env_layer(grid = grid, n = 3)$dataframe |> dplyr::rename(lon = x, lat = y)
str(generate_resource_layer(env_layers = cdt2, coordinate_fields = c("lon", "lat"),
                            beta = c(2, -1.5, 3)) )
}
