% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R, R/test-normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(cov)

normalize(cov)
}
\arguments{
\item{cov}{a vector}
}
\value{
a vector

a vector
}
\description{
Normalize a vector

Normalize a vector
}
\examples{
normalize(c(1:5))
normalize(c(1:5))
}
