% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_position_func.R
\name{potential_position_func}
\alias{potential_position_func}
\title{Identify potential next positions}
\usage{
potential_position_func(
  n,
  bearing,
  step,
  from,
  colony_location,
  resource_layer
)
}
\arguments{
\item{n}{Numeric, the number of potential positions to generate}

\item{bearing}{Numeric vector, mu and kappa to be passed on randomdir}

\item{step}{Numeric, shape and rate to be passed on randomdist}

\item{from}{data.frame including Lon and Lat fields and stepID of the point of origin}

\item{colony_location}{data.frame vector with lon and lat position of the colony}

\item{resource.layer}{SpatRaster, the suitability or any resource layer to be used to sort potential position (a single layer)}
}
\value{
A data.frame with the n potential positions from the point of origin, informed with the value of the provided environmental layer corresponding to each as well as the distance from the colony.
}
\description{
Identify potential next positions
}
\examples{
from <- data.frame(Lon = 1, Lat = 2, toto = 5)
colony_location <- data.frame(Lon = 2, Lat = 5)
cdt <- generate_env_layer(grid = create_grid(), n = 1)$rasters
potential_position_func(n = 10, from = from, 
                        colony_location = colony_location, resource_layer = cdt, 
                        bearing = c(90,10), step = c(4.5, 3))
}
