% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize.R
\name{CreateSegment}
\alias{CreateSegment}
\alias{CreateSegments}
\alias{MergeLast}
\alias{SegmentSpatialLines}
\alias{Merge}
\title{Author: Auriane Virgili}
\usage{
CreateSegment(coords, from, to)

CreateSegments(coords, length = 0, n.parts = 0)

MergeLast(lst)

SegmentSpatialLines(sl, length = 0, n.parts = 0, merge.last = FALSE)

Merge(segments)
}
\arguments{
\item{coords}{}

\item{from}{}

\item{to}{}

\item{length}{length to segmentize}

\item{n.parts}{number of parts to cut the transect into}

\item{lst}{}

\item{sl}{a spatial line object}

\item{merge.last}{should the last bit be merged with the previous one?}

\item{segments}{a sf object}
}
\value{
a sf object
}
\description{
Author: Auriane Virgili
}
