% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createsegment.R
\name{SegmentSpatialLines}
\alias{SegmentSpatialLines}
\title{Title}
\usage{
SegmentSpatialLines(sl, length = 0, n.parts = 0, merge.last = FALSE)
}
\arguments{
\item{sl}{a spatial line object}

\item{length}{length to segmentize}

\item{n.parts}{number of parts to cut the transect into}

\item{merge.last}{should the last bit be merged with the previous one?}
}
\description{
Title
}
