% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtools-package.R
\docType{package}
\name{simtools-package}
\alias{simtools}
\alias{simtools-package}
\title{simtools: Tools For Simulation Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Includes a suite of tools to simulate virtual ecological systems, from environment, to animals and surveys.
}
\author{
\strong{Maintainer}: Charlotte Lambert \email{charlotte.lambert@univ-lr.fr} (\href{https://orcid.org/0000-0002-1128-5262}{ORCID})

}
\keyword{internal}
