% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_spat.R
\name{simul_spat}
\alias{simul_spat}
\title{Simulate spatial point process}
\usage{
simul_spat(
  ref_map,
  N,
  seed = NULL,
  n_sim = 1,
  return_wgs_coordinates = TRUE,
  drop_geometry = FALSE,
  mean_group_size = NULL
)
}
\arguments{
\item{ref_map}{Reference map, must be projected}

\item{N}{Number of individuals in the final population}

\item{seed}{Numeric}

\item{n_sim}{Number of simulation to run}

\item{return_wgs_coordinates}{Boolean. Should the function returns points in WGS coordinates? Recommended when working on real environment}

\item{drop_geometry}{Boolean. Should the geometry be dropped in returned object? Defaults to FALSE, if TRUE the function will return a data.frame and not a sf object.}

\item{mean_group_size}{The mean group size to generate the mark. Set to NULL if no mark is wanted.}
}
\value{
Returns an sf object
}
\description{
Generates a spatial point process from a reference map, using Inhomogeneous Poisson Point Process with spatstat IPP. The intensity of the process is proportional to this reference map, and the final number of individuals in the population is provided by N. Several runs can be simulated at the same time. If working in a real environment, reference map must be projected.
}
\examples{
grid <- create_grid()
cdt <- generate_env_layer(grid = grid)
rsce <- generate_resource_layer(env_layers = cdt$rasters,
                            beta = c(2, -1.5))

# issue a warning: we are in a virtual env and ref_map has no crs
out <- simul_spat(ref_map = rsce$rasters, n_sim = 2, 
                  return_wgs_coordinates = FALSE, N = 1000, mean_group_size = 10)

ggplot2::ggplot(out) +
  ggplot2::geom_tile(data = rsce$dataframe, ggplot2::aes(x = x, y = y, fill = suitability)) + 
  ggplot2::geom_sf() + ggplot2::facet_wrap("name") +
  viridis::scale_fill_viridis(option = "H")

ggplot2::ggplot(out) +
   ggplot2::geom_sf(ggplot2::aes(size = size)) + ggplot2::facet_wrap("name") 

}
\seealso{
\code{\link{generate_resource_layer()}} \code{\link{generate_env_layers()}}
}
