% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomdir.R
\name{randomdir}
\alias{randomdir}
\title{Return random direction angles}
\usage{
randomdir(n, mu, kappa)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean direction of the distribution as a \code{circular} object.}
}
\description{
Return random direction angles
}
