% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomdir.R
\name{randomdir}
\alias{randomdir}
\title{Return random direction angles}
\usage{
randomdir(n, mu, kappa)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean direction of the distribution as a \code{circular} object.}

\item{kappa}{numeric, sd parameter}
}
\value{
n random angles
}
\description{
Return random direction angles
}
\seealso{
Other individual movement functions: 
\code{\link{potential_position_func}()},
\code{\link{randomdist}()},
\code{\link{simulate_trajectory_CPF}()},
\code{\link{simulate_trajectory_FR}()}
}
\concept{individual movement functions}
